Organize and rename files
Created a src/ folder for keeping python files.
Renamed FileTest.py to test_File.py
Fixed import statement that used old test module name.
